ingestion/requirements.txt
kafka-python
pandas
astroquery
